date_14_june_25

+++ src/main/java/net/harshDeveloper/JournalApp/services/JounalEntryService.java
@@ -6,6 +6,7 @@ import net.harshDeveloper.JournalApp.repository.JounalEntryRepositopry;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -24,13 +25,19 @@ public class JounalEntryService {
 
 
 
-    public  void SaveEntry(JounalEntry jounalEntry , String username){
-
-        User user =  userService.findByUsername(username) ;
+    @Transactional
+    JounalEntry SaveEntry(JounalEntry jounalEntry , String username){
+    try {
+        User user = userService.findByUsername(username);
         jounalEntry.setdate(LocalDateTime.now());
         JounalEntry saved = jounalEntryRepository.save(jounalEntry);
         user.getJounalEntries().add(saved); //System.out.println("→ [Service] saved with ID: " + saved.getId());
-       userService.saveUser(user);
+        userService.saveUser(user);
+        return saved;
+    } catch (Exception e) {
+        System.out.println(e);
+        throw new RuntimeException("An error occur while saving the entry. " , e);
+    }
     }
 
 
@@ -43,9 +50,19 @@ public class JounalEntryService {
     }
 
 
-    public JounalEntry deleteElemetbyId(ObjectId myid){
-        jounalEntryRepository.deleteById(myid);
-        return  null;
+
+    public void deleteElemetbyId(ObjectId myid) {
+        Optional<JounalEntry> entryOptional = jounalEntryRepository.findById(myid);
+        if (entryOptional.isPresent()) {
+            JounalEntry entry = entryOptional.get();
+            User user = entry.getUser();
+
+            if (user != null) {
+                user.removeJournalEntry(entry); // Remove from user's list
+                userService.saveUser(user);
+            }
+            jounalEntryRepository.deleteById(myid);
+        }
     }
 
 }
